+++ src/employee/employee.controller.js
@@ -30,7 +30,7 @@ class EmployeeController {
 
     async deleteEmployee(req, res) {
         try {
-            const data = await employeeService.deleteEmployee({ id: req.params.id })
+            const data = await employeeService.deleteEmployee({ id: req.params.id, userId: req.query.id, ...req.body })
             res.status(200).json(data)
         } catch (error) {
             res.status(500).json({ message: error.message })
